Fase 2/Evidencias Proyecto/Evidencias de sistema/Backend/backend/app/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
